Add files via upload
This are the 3d obj created in tinker cad. This is the 3d map please extract it where obj1.zip and obj.zip files will be gathered and inserted in the main directory while Departure.zip has the obj and mtl for departure which will be the Departure folder is inserted inside the main directory.